added options for composer php
The version option was added due to suspicions that a faulty dev
deployment of tag1consulting.com was caused by composer being ran on
php7.3, while the site is ran on php5.6.

The suspicions were raised when looking at the apache error logs. This
was the most noticeable message:

[Mon Aug 26 23:04:19.929822 2019] [proxy_fcgi:error] [pid 23017:tid
140565848991488] [client 172.31.24.205:51780] AH01071: Got error 'PHP
message: PHP Parse error:  syntax error, unexpected ':', expecting ';'
or '{' in
/var/www/dev.tag1consulting.com/releases/20180331/vendor/doctrine/annotations/lib/Doctrine/Common/Annotations/AnnotationRegistry.php
on line 50\n'

The syntax error was on the new php7 type hinting syntax, meaning
composer pulled the wrong version of the dependency.

The php_version option is going to be passed on the command line when
ran from jenkins, etc.